refactor: remove stores and v1 api endpoints & fix tests
