+++ operate/client/src/App/Processes/ListView/InstancesTable/Toolbar/tests/mocks.tsx
@@ -10,7 +10,7 @@ import {useEffect} from 'react';
 import {MemoryRouter} from 'react-router-dom';
 import {QueryClientProvider} from '@tanstack/react-query';
 import {getMockQueryClient} from 'modules/react-query/mockQueryClient';
-import {processInstancesSelectionStore} from 'modules/stores/processInstancesSelectionV2';
+import {processInstancesSelectionStore} from 'modules/stores/processInstancesSelection';
 import {batchModificationStore} from 'modules/stores/batchModification';
 import type {ProcessInstance} from '@camunda/camunda-api-zod-schemas/8.8';
 import {createProcessDefinition} from 'modules/testUtils';
