+++ operate/client/src/App/Processes/ListView/tests/mocks.tsx
@@ -8,7 +8,7 @@
 
 import {MemoryRouter} from 'react-router-dom';
 import {useEffect} from 'react';
-import {processInstancesSelectionStore} from 'modules/stores/processInstancesSelectionV2';
+import {processInstancesSelectionStore} from 'modules/stores/processInstancesSelection';
 import {operationsStore} from 'modules/stores/operations';
 import {batchModificationStore} from 'modules/stores/batchModification';
 import {Paths} from 'modules/Routes';
