+++ operate/client/src/App/Processes/MigrationView/TopPanel/Diagrams/v2/SourceDiagram.test.tsx
@@ -14,7 +14,7 @@ import {QueryClientProvider} from '@tanstack/react-query';
 import {getMockQueryClient} from 'modules/react-query/mockQueryClient';
 import {processInstanceMigrationStore} from 'modules/stores/processInstanceMigration';
 import {useEffect} from 'react';
-import {processInstancesSelectionStore} from 'modules/stores/processInstancesSelectionV2';
+import {processInstancesSelectionStore} from 'modules/stores/processInstancesSelection';
 import {mockFetchProcessDefinitionXml} from 'modules/mocks/api/v2/processDefinitions/fetchProcessDefinitionXml';
 import * as filterModule from 'modules/hooks/useProcessInstanceStatisticsFilters';
 import {MemoryRouter} from 'react-router-dom';
