+++ operate/client/src/modules/hooks/useBatchOperationMutationRequestBody.ts
@@ -11,13 +11,6 @@ import {variableFilterStore} from 'modules/stores/variableFilter';
 import {processInstancesSelectionStore} from 'modules/stores/processInstancesSelection';
 import {buildMutationRequestBody} from 'modules/utils/buildMutationRequestBody';
 
-/**
- * A hook that builds the request body for process instance batch operations,
- * taking into account:
- * - current search parameters from the URL
- * - selected and excluded IDs from the processInstancesSelectionStore
- * - variable filters from variableFilterStore.
- */
 const useBatchOperationMutationRequestBody = () => {
   const variable = variableFilterStore.variable;
   const [searchParams] = useSearchParams();
